+++ lib/src/constants.dart
@@ -1,0 +1,2 @@
+// Default letter height ratio determined from unspecified font family.
+const double defaultLetterHeightRatio = 2536.0 / 3600.0;